pyproject.toml
[tool.poetry]
name = "pacilfess-discord"
version = "0.1.0"
description = ""
authors = ["Rendy Arya Kemal <renrror@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
discord-py-slash-command = "^2.3.2"
aiosqlite = "^0.17.0"
dataclasses-json = "^0.5.4"
prettytable = "^2.1.0"
pycryptodome = "^3.10.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.7b0"
isort = "^5.9.2"
flake8 = "^3.9.2"
mypy = "^0.910"
"discord.py-stubs" = "^1.7.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
